'短信验证码发送及职员身份证密码登陆完成'

+++ src/main/java/com/yuyi/tea/service/RedisService.java
@@ -51,7 +51,7 @@ public class RedisService {
      * @param value
      * @return
      */
-    public boolean set(final String key, Object value, Long expireTime, TimeUnit timeUnit) {
+    public boolean set(final String key, Object value, int expireTime, TimeUnit timeUnit) {
         boolean result = false;
         try {
             ValueOperations<Serializable, Object> operations = redisTemplate.opsForValue();
